+++ src/Exception/ValidationFailedExceptionInterface.php
@@ -1,29 +1,34 @@
 <?php
+declare(strict_types=1);
 
 namespace Dhii\Validation\Exception;
 
 use Dhii\Util\String\StringableInterface as Stringable;
-use Traversable;
-use Dhii\Validation\SubjectAwareInterface;
+use Exception;
 
 /**
  * Something that can represent an exception which occurs if a validation
  * fails.
- *
- * @since 0.1
  */
-interface ValidationFailedExceptionInterface extends
-        ValidationExceptionInterface,
-        SubjectAwareInterface
+interface ValidationFailedExceptionInterface extends ValidationExceptionInterface
 {
     /**
      * Retrieves validation errors that are associated with this instance.
      *
-     * @since 0.1
+     * @return string[]|Stringable[]|iterable A list of errors.
+     *                                        Each error is something that can be treated as a string, and represents
+     *                                        a description of why a validation subject is invalid.
      *
-     * @return string[]|Stringable[]|Traversable A list of errors.
-     *                                           Each error is something that can be treated as a string, and represents
-     *                                           a description of why a validation subject is invalid.
+     * @throws Exception If problem retrieving.
      */
-    public function getValidationErrors();
+    public function getValidationErrors(): iterable;
+
+    /**
+     * Returns the subject, the validation for which failed.
+     *
+     * @return mixed The subject that was being validated.
+     *
+     * @throws Exception If problem retrieving.
+     */
+    public function getValidationSubject();
 }