+++ src/ValidatorInterface.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Dhii\Validation;
 
@@ -7,15 +8,11 @@ use Dhii\Validation\Exception\ValidationFailedExceptionInterface;
 
 /**
  * Something that can validate a value.
- *
- * @since 0.1
  */
 interface ValidatorInterface
 {
     /**
      * Validates a value.
-     *
-     * @since 0.1
      *
      * @param mixed $value The subject of validation.
      *