+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [[*next-version*]] - YYYY-MM-DD
+### Removed
+- `SpecAwareInterface`.
+- `ValidatorFactoryInterface`.
+- `SubjectAwareInterface`.
+- `ValidatorAwareInterface`.
+- `@since` tags everywhere.
+- Support for PHP < 7.1.
+
+### Changed
+- `ValidationFailedExceptionInterface::getSubject()` is now `getValidationSubject().
+
 ## [0.2] - 2018-08-29
 Stable release. No code changed.